refactor: refactoring after code review
