+++ migration-newip-to-ts/src/components/view/sources/sources.ts
@@ -1,18 +1,5 @@
 import './sources.css';
-
-export interface SourcesInterface {
-    draw(data: ISources[]): void;
-}
-
-export type ISources = {
-    id: string;
-    name: string;
-    description: string;
-    category: string;
-    language: string;
-    country: string;
-    url: string;
-};
+import { ISources, SourcesInterface } from '../../../types/sources';
 
 class Sources implements SourcesInterface {
     draw(data: ISources[]) {
@@ -23,7 +10,9 @@ class Sources implements SourcesInterface {
             if (sourceItemTemp) {
                 const sourceClone = <HTMLTemplateElement>sourceItemTemp.content.cloneNode(true);
                 const sourceItemName = sourceClone.querySelector('.source__item-name');
-                if (sourceItemName) sourceItemName.textContent = item.name;
+                if (sourceItemName) {
+                    sourceItemName.textContent = item.name;
+                }
                 sourceClone.querySelector('.source__item')?.setAttribute('data-source-id', item.id);
 
                 fragment.append(sourceClone);
@@ -34,8 +23,10 @@ class Sources implements SourcesInterface {
         sources?.append(fragment);
 
         const firstSource: HTMLElement | null | undefined = <HTMLElement>sources?.firstElementChild;
-        if (firstSource) firstSource.click();
-        firstSource.classList.add('active');
+        if (firstSource) {
+            firstSource.click();
+            firstSource.classList.add('active');
+        }
     }
 }
