+++ migration-newip-to-ts/tsconfig.json
@@ -6,6 +6,7 @@
     "module": "es6",
     "target": "es5",
     "allowJs": false,
+    "moduleResolution": "Node",
     "sourceMap": true
   }
 }
